requirements.txt
# Core ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
numpy>=1.24.0
pandas>=2.0.0

# Experiment tracking
mlflow>=2.10.0

# App
streamlit>=1.32.0
openai>=1.12.0

# Drift detection
evidently>=0.7.0

# Utilities
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
